Handle an error by kind, or propagate it: nine sites where a failure became a plausible value
One rule, applied where the value reaches a table or a key. `visit` in `tree.rs` is
the pattern the crate already had -- `NotFound` only, everything else `?`, with a
comment explaining why swallowing more would let the store validate a truncated
tree.

- `report::read_json` returned `Option` and collapsed ABSENT with UNREADABLE, and
  two bare `continue`s dropped the row. A truncated `summary.json` -- `fs::write`
  truncates before writing, so a SIGKILLed sweep leaves one -- was indistinguishable
  from a battery never scored. The three invariants below cannot see it: (1) and (1b)
  compare MAXIMA per agent and per battery, so another tool still supplies the max,
  and (1c) only checks the run's directory exists. Its sibling
  `generate_harvest_bench` documents exactly this: "a missing record dropped its row
  ... the table published N-1 rows of N and announced success". An attested battery
  with no record now refuses by name. `aggregate_cases`, `kiro_cost` and
  `generate_numbers_tex` become fallible with it.

- `runtests`' `extract` collapsed "regex matched nothing" to `0` via `unwrap_or(0)`.
  `measured_nothing` and `reconcile` guard the CASE counts -- a zeroed
  `cases_discovered` fails reconciliation -- while the three VECTOR counts had
  nothing checking them at all, so a reworded summary line in the vendored
  test-corpus submodule would publish `0/0` vectors and regenerate `tables/` from it.
  Now each named line must be present and parse.

- `build_harvest_bench_lib` never inspected `out.status`: the verdict was
  `so.is_file()` alone. A missing `timeout` or `cargo` on the scoring process's PATH,
  `timeout`'s 124, or a fork returning EAGAIN under `--parallel` printed "build
  failed (no cdylib)" and recorded `Builds: no` for all seven projects at once --
  which is what a cmake needing 3.24 on a 3.22 box already did to seven crates that
  compiled perfectly. 124 and 127 are now named as infra.

- `count_unsafe` treated a file `syn` cannot parse as ZERO unsafe, while `count_loc`
  -- a separate walk that never parses -- still counted every one of its lines. A
  crate full of `unsafe` with one unbalanced delimiter published a low Unsafe
  percentage against its full LOC. Not hypothetical: `Enrichment::compute` runs on
  crates that FAILED TO BUILD, and runtests records 76 build failures across the
  corpus. `UnsafeCounts::unparsed_files` records the gap instead.

- `battery::build_config` used `unwrap_or_default()` on `extract_features` inside a
  function already returning `Result`. `extract_features_from_path` already returns
  `Ok(vec![])` for an ABSENT presets file, so the only thing this hid was a MALFORMED
  one -- and then `resolve_features` yields nothing, the `if !resolved.is_empty()`
  guard skips `set_default_features`, and the follower keeps the LEADER's features:
  built as one configuration, graded against another's vectors.

- `io::workdir::resolve` did `read_to_string("/proc/mounts").unwrap_or_default()`.
  Empty means `fstype_for` matches no prefix and returns `None`, so the tmpfs refusal
  is SKIPPED -- on exactly the hosts the module exists for, since `ENV_ALLOW_TMPFS`
  is documented as the escape hatch "for hosts whose only writable filesystem is a
  tmpfs (some CI)". Per-case `target/` trees would be charged to RAM and the cgroup
  killed with no result recorded.

- `--parallel 0` parsed, and `Semaphore { max: 0 }` made `acquire` wait on a condvar
  that only a guard could notify, with no guard able to exist: the sweep hung
  silently. Refused at the flag now, via a `value_parser` range.

- `assert_pins_honoured` swallowed every read error as "nothing to compare"; only
  `NotFound` does now.

One gate changed, deliberately, and it is the second of CLAUDE.md's three moves --
give the check the information to judge correctly.
`only_the_pool_mints_the_runs_concurrency_budget` was a bare `contains(
"Semaphore::new(")` over file text, so the doc comment explaining why
`Semaphore::new(0)` deadlocks counted as minting one. It now strips line comments
first. The cheap way out was to reword the comment, which leaves the next reader to
trip over it again. Verified still non-vacuous: a planted `const _PLANTED: &str =
"Semaphore::new(1)"` in `eval.rs` fails it.

+++ tools/src/cli.rs
@@ -128,11 +128,15 @@ pub struct Cli {
     /// Beside `--tool` because it is that list's budget: with three tools this is three in flight
     /// each, nine in total. `global`, so either position parses -- without it `run HB --parallel 3` is
     /// a clap error, which is how a sweep launcher died on `unexpected argument`.
-    #[arg(long, global = true, default_value_t = 1)]
-    pub parallel: usize,
+    /// `0` is refused at the flag rather than deadlocking. `Semaphore::new(0)` sets `max = 0`, so
+    /// `Pool::acquire` loops `while 0 >= 0` on a condvar only a guard could notify, and no guard can
+    /// ever exist: the sweep hung silently instead of saying what was wrong.
+    #[arg(long, global = true, default_value_t = 1, value_parser = clap::value_parser!(u16).range(1..))]
+    pub parallel: u16,
 
-    /// Model id. Required for `--tool oneshot` and `opencode`, where the model IS the
-    /// identity; fixed by the tool otherwise.
+    /// Model id. `--tool codex` accepts one to reach the historical models; claude's comes from
+    /// `HARVEST_CLAUDE_MODEL` or its default, and kiro's is fixed. Every tool pins one either way --
+    /// there is no longer any way to run without a model in the key.
     #[arg(long)]
     pub model: Option<String>,
